Refactoring graph splitter
