+++ federatedscope/core/splitters/graph/__init__.py
@@ -1,0 +1,20 @@
+from __future__ import absolute_import
+from __future__ import print_function
+from __future__ import division
+
+from federatedscope.core.splitters.graph.louvain_splitter import LouvainSplitter
+from federatedscope.core.splitters.graph.random_splitter import RandomSplitter
+
+from federatedscope.core.splitters.graph.reltype_splitter import RelTypeSplitter
+
+from federatedscope.core.splitters.graph.scaffold_splitter import ScaffoldSplitter
+from federatedscope.core.splitters.graph.graphtype_splitter import GraphTypeSplitter
+from federatedscope.core.splitters.graph.randchunk_splitter import RandChunkSplitter
+
+from federatedscope.core.splitters.graph.analyzer import Analyzer
+
+
+__all__ = [
+    'LouvainSplitter', 'RandomSplitter', 'RelTypeSplitter', 'ScaffoldSplitter',
+    'GraphTypeSplitter', 'RandChunkSplitter', 'Analyzer'
+]
