+++ federatedscope/core/splitters/graph/randchunk_splitter.py
@@ -1,4 +1,3 @@
-import torch
 import numpy as np
