--- federatedscope/gfl/dataset/splitter/__init__.py
@@ -1,16 +1,0 @@
-from __future__ import absolute_import
-from __future__ import print_function
-from __future__ import division
-
-from federatedscope.gfl.dataset.splitter.louvain_splitter import LouvainSplitter
-from federatedscope.gfl.dataset.splitter.random_splitter import RandomSplitter
-from federatedscope.gfl.dataset.splitter.reltype_splitter import RelTypeSplitter
-from federatedscope.gfl.dataset.splitter.graphtype_splitter import GraphTypeSplitter
-from federatedscope.gfl.dataset.splitter.scaffold_splitter import ScaffoldSplitter
-from federatedscope.gfl.dataset.splitter.randchunk_splitter import RandChunkSplitter
-from federatedscope.gfl.dataset.splitter.analyzer import Analyzer
-
-__all__ = [
-    'LouvainSplitter', 'RandomSplitter', 'RelTypeSplitter', 'ScaffoldSplitter',
-    'GraphTypeSplitter', 'randchunk_splitter', 'Analyzer'
-]
